refactor: reorganize control and peripheral modules
